LUM-13 -code refactor
