+++ apps/api/src/modules/auth/auth.schema.ts
@@ -1,7 +1,6 @@
 import type { TypeOf } from "zod";
+import { selectUserSchema } from "node_modules/@lumpik/db/src/tables";
 import { object, string, void as zodVoid } from "zod";
-
-import { selectUserSchema } from "../../../../../packages/db/src/tables";
 
 export const createUserSchema = {
   input: object({
