+++ apps/api/src/utils/logger.ts
@@ -1,6 +1,6 @@
 import pino from "pino";
 
-import { env } from "../plugins/config/env";
-import { loggerConfig } from "../plugins/config/logger.config";
+import { env } from "../config/env";
+import { loggerConfig } from "../config/logger.config";
 
 export const logger = pino(loggerConfig[env.ENV]);
